login comparisons complete

+++ src/main/java/edu/xavier/csci/flowerpower/User.java
@@ -74,7 +74,8 @@ public class User {
         this.email = email;
     }
 
-    public long getId() {return id;}
+    public int getId() {return id;}
     public void setId(int id){this.id = id;}
+
 
 }
